Move Integration tests to separate package under tests
This is the first stage of improving the way how integration
tests are run in our CI - first we want to separate them in
a separate packages, and then we want to run them separately - one
integration each in the CI.
