+++ tests/cli/commands/test_celery_command.py
@@ -61,37 +61,6 @@ class TestWorkerPrecheck:
         assert airflow.settings.validate_session() is False
 
 
-@pytest.mark.integration("redis")
-@pytest.mark.integration("rabbitmq")
-@pytest.mark.backend("mysql", "postgres")
-class TestWorkerServeLogs:
-    @classmethod
-    def setup_class(cls):
-        cls.parser = cli_parser.get_parser()
-
-    @mock.patch("airflow.cli.commands.celery_command.celery_app")
-    @conf_vars({("core", "executor"): "CeleryExecutor"})
-    def test_serve_logs_on_worker_start(self, mock_celery_app):
-        with mock.patch("airflow.cli.commands.celery_command.Process") as mock_process:
-            args = self.parser.parse_args(["celery", "worker", "--concurrency", "1"])
-
-            with mock.patch("celery.platforms.check_privileges") as mock_privil:
-                mock_privil.return_value = 0
-                celery_command.worker(args)
-                mock_process.assert_called()
-
-    @mock.patch("airflow.cli.commands.celery_command.celery_app")
-    @conf_vars({("core", "executor"): "CeleryExecutor"})
-    def test_skip_serve_logs_on_worker_start(self, mock_celery_app):
-        with mock.patch("airflow.cli.commands.celery_command.Process") as mock_popen:
-            args = self.parser.parse_args(["celery", "worker", "--concurrency", "1", "--skip-serve-logs"])
-
-            with mock.patch("celery.platforms.check_privileges") as mock_privil:
-                mock_privil.return_value = 0
-                celery_command.worker(args)
-                mock_popen.assert_not_called()
-
-
 @pytest.mark.backend("mysql", "postgres")
 class TestCeleryStopCommand:
     @classmethod
