+++ tests/integration/providers/redis/operators/test_redis_publish.py
