+++ tests/integration/providers/redis/sensors/test_redis_key.py
